Refactoring the email/SMS reminder system so it's easier to change up what emails get sent when, but making it easy to reuse common emails and templates whenever possible.
